Okay lot of things learned today.
Lot of interchangeable stuff goign on I guess. What works vs what doesnt work. More amusing to play with stuff it seems.

But this is a pretty simple search so far.
I want to set this to a grid or layout view. I dunno. I havent decided that when I search up top I'll have the results down below. Undecided.